requirements.txt
# ── LangChain stack ───────────────────────────────────────────────────────────
langchain
langchain-classic
langchain-google-vertexai
langchain-postgres
langchain-community
langchain-text-splitters

# ── LlamaIndex stack ──────────────────────────────────────────────────────────
llama-index-core
llama-index-llms-vertex
llama-index-embeddings-vertex
llama-index-vector-stores-postgres

# ── API ─────────────────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]
python-multipart

# ── Shared ────────────────────────────────────────────────────────────────────
psycopg[binary]==3.2.6
pgvector==0.3.6
SQLAlchemy==2.0.49
python-dotenv==1.1.0
pypdf==5.4.0
